requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
Pillow==11.1.0
python-dotenv==1.0.1
Requests==2.32.3
ttkbootstrap==1.10.1
tvdb_v4_official==1.1.0
